+++ ArduPlane/Parameters.pde
@@ -609,7 +609,7 @@ static void load_parameters(void)
 		g.format_version.set_and_save(Parameters::k_format_version);
 		Serial.println_P(PSTR("done."));
     } else {
-	    unsigned long before = micros();
+	    uint32_t before = micros();
 	    // Load all auto-loaded EEPROM variables
 	    AP_Param::load_all();
